+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Added the option to clone a story.
 
+### Changed
+- Moved the story action state buttons to a react component.
+
 ## [1.14.0] 2017-10-18
 ### Added
 - Added adminer service to docker-compose.